Dockerize finished. Nginx,laravel app and mysql added in docker compose. Related tests and ratelimit test added. SSL added for webserver. LocationRequests revised for marker color validation.

+++ database/migrations/2024_10_31_094751_create_locations_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('created_user_id')->constrained()->on('users')->cascadeOnDelete();
             $table->string('name');
-            $table->integer('lat');
-            $table->integer('long');
+            $table->double('lat');
+            $table->double('long');
             $table->string('marker_color');
             $table->timestamps();
         });
